add conftest.py and fix tests
